Consolidated string helper functions
