--- _strcmp.c
@@ -1,14 +1,0 @@
-#include "stringwrapper.h"
-
-/**
- * _strcmp - a function that compares two strings.
- * @s1: first string
- * @s2: second string
- * Return: value of comparison
- */
-int _strcmp(const char *s1, const char *s2)
-{
-	while (*s1 && (*s1 == *s2))
-		s1++, s2++;
-	return (*(const unsigned char *)s1 - *(const unsigned char *)s2);
-}
